Avoid numbers in enums for Swift compatibility
Since `WPAnalyticsStat` is the name of the enum, Swift will remove that
as a prefix so you can do things like `.LoginFailed`.

In #36, a number of 3D Touch related constants were introduced, which
would be mapped as `.3DTouchShortcut*` enum values. As I understand it,
Swift names can't start with a digit, which is probably why the compiler
was only stripping the `WPAnalytics` prefix but leaving the `Stats`, and
all the enum references in Swift were broken.

I know it's not really called Force Touch on the iPhone, so feel free to
use an alternative, as long as it doesn't start with a digit.